B other (movie): FntFlush(long)->FntFlush(int) to match real PsyQ libgpu -> ccpsx clean
movie_externs.h:30 declared FntFlush(long id), conflicting with real PsyQ libgpu.h
FntFlush(int id). Matched signature -> identical (non-conflicting) redeclaration.
RC=0 both gates. (ResetCallback/VSync implicit-decl remain as benign warnings -
libetc not yet in ccpsx include set; candidate for Step-3 include expansion.)

+++ frontend/psx/movie_externs.h
@@ -27,7 +27,7 @@ void DecDCTout(u_long * cell, long size);
 long DecDCToutCallback(void * func);
 int DecDCTvlc(u_long * bs, u_long * buf);
 int DecDCTvlcSize(int size);
-u_long * FntFlush(long id);
+u_long * FntFlush(int id);   /* match real PsyQ libgpu.h proto (int, not long) */
 void Platform_ResetDCTBuffer(void);
 int SNDcdvol(int vol);
 void StCdInterrupt(void);
